Moved all method implementations into the concrete subclass, and removed the default methods in interfaces (and banned them via checkstyle).
